Refactored code into separate files (re-tested) Rubocop
