+++ lib/tasks/facilities_management/postcode.rake
@@ -1,6 +1,7 @@
 require 'optparse'
 require Rails.root.join('lib', 'tasks', 'ordnance_survey')
 
+# rubocop:disable Metrics/BlockLength
 namespace :db do
   desc 'add OS postcode data to the database'
   # task :postcode, [:access_key, :secret_access_key, :bucket, :region] => :environment do |_, args|
@@ -10,7 +11,7 @@ namespace :db do
     OrdnanceSurvey.create_address_lookup_view
     OrdnanceSurvey.create_postcode_locator_index
     OrdnanceSurvey.create_new_postcode_views
-    OrdnanceSurvey.import_postcodes args[:access_key], args[:secret_access_key], args[:bucket], args[:region]
+    OrdnanceSurvey.import_postcodes 'dataPostcode2files', args[:access_key], args[:secret_access_key], args[:bucket], args[:region]
   end
 
   task :local_postcode, [:folder] => :environment do |_, args|
@@ -22,7 +23,7 @@ namespace :db do
     end
     nargs = o.order!(ARGV) {}
     o.parse!(nargs)
-    options[:folder] = args[:folder] if options.empty?  # support debugging
+    options[:folder] = args[:folder] if options.empty? # support debugging
 
     p "Creating postcode database and local import from #{(options[:folder] || Rails.root.join('data', 'local_postcodes'))}"
     Rails.logger.info "Creating postcode database and local import from #{(options[:folder] || Rails.root.join('data', 'local_postcodes'))}"
@@ -98,3 +99,4 @@ namespace :db do
     OrdnanceSurvey.create_new_postcode_views
   end
 end
+# rubocop:enable Metrics/BlockLength
