+++ lib/tasks/ordnance_survey.rb
@@ -10,7 +10,11 @@ module OrdnanceSurvey
   require 'aws-sdk-s3'
   require 'json'
   require Rails.root.join('lib', 'tasks', 'distributed_locks')
-  include ActiveSupport::NumberHelper
+  require Rails.root.join('lib', 'tasks', 'os_data_processing')
+  require Rails.root.join('lib', 'tasks', 'os_file_handler')
+  require Rails.root.join('lib', 'tasks', 'os_stream_handler')
+
+  extend ActiveSupport::NumberHelper
 
   def self.create_postcode_table
     str   = File.read(Rails.root + 'data/postcode/PostgreSQL_AddressBase_Plus_CreateTable.sql')
@@ -198,77 +202,10 @@ module OrdnanceSurvey
           updated_at timestamp without time zone NOT NULL);"
     ActiveRecord::Base.connection_pool.with_connection { |db| db.exec_query query }
 
-    # -- Indices -------------------------------------------------------
-    query = "
-CREATE UNIQUE INDEX IF NOT EXISTS os_address_admin_uploads_filename_idx ON os_address_admin_uploads USING btree (filename);"
+    query = 'CREATE UNIQUE INDEX IF NOT EXISTS os_address_admin_uploads_filename_idx ON os_address_admin_uploads USING btree (filename);'
     ActiveRecord::Base.connection_pool.with_connection { |db| db.exec_query query }
   rescue PG::Error => e
     puts e.message
-  end
-
-  def self.postcode_file_already_loaded(filename)
-    query = "select count(*) from os_address_admin_uploads where fail_reason is null and (filename = '#{filename}' or filename = '#{extract_metadata(filename)}')"
-
-    ActiveRecord::Base.connection_pool.with_connection do |db|
-      result = db.exec_query(query)
-      count  = result[0]['count']
-      p "Skipping file #{filename}" if count.positive?
-      return true if count.positive?
-    end
-    false
-  rescue PG::Error => e
-    puts e.message
-    false
-  end
-
-  MATCH_1 = /dataPostcode_(?<meta>(?<date>\d*-\d*-\d*)_(?<seq>\d*)_(?<outcode>[\w]{1,2}))/.freeze
-  MATCH_2 = /AddressBasePlus_.*?_(?<meta>(?<date>\d*-\d*-\d*)_(?<seq>\d*)((?>.csv-)(?<outcode>\w{1,2}))?)/.freeze
-
-  def self.extract_metadata(filename, outcode = nil)
-    [MATCH_1, MATCH_2].each do |m|
-      match_data = filename.match(m)
-      next if match_data.nil?
-
-      result = "#{match_data[:date]}_#{match_data[:seq]}"
-      result += "_#{match_data[:outcode].upcase}" unless match_data[:outcode].nil?
-      outcode << match_data[:outcode] unless match_data[:outcode].nil? || outcode.nil?
-
-      return result
-    end
-
-    filename
-  end
-
-  def self.log_postcode_file_failed(filename, reason)
-    metadata = extract_metadata(filename)
-    reason = reason.sub('`', '').sub('\'', '').sub('/', '/')
-    query = "INSERT INTO os_address_admin_uploads (filename, fail_reason, created_at, updated_at) VALUES('#{metadata}', '#{reason}', '#{DateTime.now.utc}', '#{DateTime.now.utc}')
-    on conflict (filename)
-    do update set fail_reason = EXCLUDED.fail_reason, updated_at = EXCLUDED.updated_at;"
-    ActiveRecord::Base.connection_pool.with_connection do |db|
-      db.exec_query(query)
-    end
-    true
-  rescue PG::Error => e
-    puts e.message
-    Rails.logger.error(e.message)
-    false
-  end
-
-  def self.log_postcode_file_loaded(filename, size, etag, created_at, updated_at)
-    key = extract_metadata(filename)
-    query = "
-    INSERT INTO os_address_admin_uploads (filename, size, etag, created_at, updated_at) VALUES('#{key}', #{size}, '#{etag}', '#{created_at}', '#{updated_at}')
-    on conflict (filename)
-    do update set fail_reason = null, size= EXCLUDED.size, etag = EXCLUDED.etag, created_at = EXCLUDED.created_at, updated_at = EXCLUDED.updated_at;"
-    ActiveRecord::Base.connection_pool.with_connection do |db|
-      db.exec_query(query)
-    end
-    true
-  rescue PG::Error => e
-    puts e.message
-    Rails.logger.error(e.message)
-    false
   end
 
   def self.awd_credentials(access_key, secret_key, bucket, region)
@@ -284,216 +221,16 @@ CREATE UNIQUE INDEX IF NOT EXISTS os_address_admin_uploads_filename_idx ON os_ad
     Aws.config[:http_idle_timeout] = 6000
   end
 
+  # rubocop:disable Style/ClassVars
   def self.os_address_headers
     @@os_address_headers ||= File.read(Rails.root.join('data', 'postcode', 'os_address_headers.csv'))
   end
+  # rubocop:enable Style/ClassVars
 
   def self.file_type(filename)
     return :dat if filename.include? '.dat'
 
     :csv if filename.include? '.csv'
-  end
-
-  def self.untar_file(filename, summary, &block)
-    summary[:updated_time] = File.mtime(filename)
-    Gem::Package::TarReader.new(Zlib::GzipReader.open(filename)) do |tar|
-      handle_tar_contents(tar, summary, &block)
-    end
-  end
-
-  def self.untar_stream(url, summary, &block)
-    Gem::Package::TarReader.new(Zlib::GzipReader.new(URI.open(url))) do |tar|
-      handle_tar_contents(tar, summary, &block)
-    end
-  end
-
-  def self.handle_tar_contents(tar, summary, &block)
-    tar.each do |entry|
-      next unless entry.file?
-
-      summary[:length] = entry.size
-      file_stream = StringIO.new(entry.read)
-      header_line = file_stream.readline
-      (meta_type = (if header_line.downcase.include?('uprn')
-                      :csv
-                    else
-                      :dat
-                    end)) and file_stream.rewind
-      summary[:md5] = chunk_file_data(file_stream, meta_type, &block)
-    end
-  end
-
-  def self.gunzip_file(filename, summary, &block)
-    summary[:updated_time] = File.mtime(filename)
-    Zlib::GzipReader.open(filename) do |gz|
-      handle_gzip_contents(gz, summary, &block)
-    end
-  end
-
-  def self.gunzip_url(url, summary, &block)
-    Zlib::GzipReader.open(URI.open(url)) do |gz|
-      handle_gzip_contents(gz, summary, &block)
-    end
-  end
-
-  def self.handle_gzip_contents(gz, summary, &block)
-    header_line = gz.readline
-    (meta_type = (if header_line.downcase.include?('uprn')
-                    :csv
-                  else
-                    :dat
-                  end)) and gz.rewind
-    summary[:md5] = chunk_file_data(gz, meta_type, &block)
-    summary[:length] = gz.pos
-  end
-
-  def self.unzip_file(filename, summary, &block)
-    summary[:updated_time] = File.mtime(filename)
-    Zip::InputStream.open(filename) do |io|
-      handle_zip_contents( io, summary, &block)
-    end
-  end
-
-  def self.unzip_url(url, summary, &block)
-    Zip::InputStream.open(IO.new(URI.open(url))) do |io|
-      handle_zip_contents( io, summary, &block)
-    end
-  end
-
-  def self.handle_zip_contents(io, summary, &block)
-    while (entry = io.get_next_entry)
-      next if entry.name_is_directory?
-
-      header_line = io.readline
-      (meta_type = (if header_line.downcase.include?('uprn')
-                      :csv
-                    else
-                      :dat
-                    end)) and io.rewind
-      summary[:length] = entry.size
-      summary[:md5] = chunk_file_data(io, meta_type, &block)
-    end
-  end
-
-  # rubocop:disable Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
-  def self.stream_url(obj, data_summary, &_block)
-    meta_type = :dat
-    chunk_count = -1
-    chunks = ''
-
-    obj.get do |chunk|
-      chunks << chunk
-      if chunk_count == -1
-        chunk_count = 0
-        meta_type = (chunk.downcase.include?('uprn') ? :csv : :dat) if chunk_count.zero?
-      end
-    end
-
-    inject_data(chunks) if meta_type == :dat
-    upsert_csv_data(StringIO.new(chunks)) if meta_type == :csv
-  rescue StandardError => e
-    data_summary[:fail] = e.message
-    raise e
-  end
-
-  def self.stream_file(filename, data_summary, &block)
-    data_summary[:updated_time] = File.mtime(filename)
-    data_summary[:length] = File.size(filename)
-    file_io = File.new(filename, 'r')
-    header_line = file_io.readline
-    (meta_type = (if header_line.downcase.include?('uprn')
-                    :csv
-                  else
-                    :dat
-                  end)) and file_io.rewind
-    data_summary[:md5] = chunk_file_data(file_io, meta_type, &block)
-  rescue StandardError => e
-    data_summary[:fail] = e.message
-    raise e
-  ensure
-    file_io&.try(&:close)
-  end
-
-  CHUNK_SIZE = 100000
-  def self.chunk_file_data(io, meta_type, &block)
-    counter = 0
-    chunk = ''
-    md5 = Digest::MD5.new
-    io.each do |line|
-      (counter = 0) and (chunk = '') if counter == CHUNK_SIZE
-      chunk << line
-      counter += 1
-      md5 << chunk if counter == CHUNK_SIZE
-      block.call(meta_type, StringIO.new(chunk)) and next if counter == CHUNK_SIZE
-    end
-    block.call(meta_type, StringIO.new(chunk)) unless chunk.empty?
-    md5.to_s
-  end
-  # rubocop:enable Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
-
-  def self.read_file_from_s3(object_summary, fn_process, &block)
-    data_summary = {}
-    data_summary[:length] = object_summary.size
-    data_summary[:md5] = object_summary.etag
-    data_summary[:updated_time] = object_summary.last_modified
-
-    case File.extname(object_summary.key)
-    when '.zip'
-      unzip_url(object_summary.presigned_url(:get), data_summary) do |type, stream|
-        fn_process.call(type, stream)
-      end
-    when '.gz'
-      if object_summary.key.include? 'tar'
-        untar_url(object_summary.presigned_url(:get), data_summary) do |type, stream|
-          fn_process.call(type, stream)
-        end
-      else
-        gunzip_url(object_summary.presigned_url(:get), data_summary) do |type, stream|
-          fn_process.call(type, stream)
-        end
-      end
-    when '.dat', '.csv'
-      stream_url(object_summary, data_summary) do |type, stream|
-        fn_process.call(type, stream)
-      end
-    else
-      Rails.logger.info "Postcode processing ignoring: #{object_summary.key}"
-    end
-    block.call(data_summary)
-  end
-
-  def self.read_file(filename, fn_process, &block)
-    data_summary = {}
-
-    case File.extname(filename)
-    when '.zip'
-      unzip_file(filename, data_summary) do |type, stream|
-        fn_process.call(type, stream)
-      end
-    when '.gz'
-      if filename.include? 'tar'
-        untar_file(filename, data_summary) do |type, stream|
-          fn_process.call(type, stream)
-        end
-      else
-        gunzip_file(filename, data_summary) do |type, stream|
-          fn_process.call(type, stream)
-        end
-      end
-    when '.dat', '.csv'
-      stream_file(filename, data_summary) do |type, stream|
-        fn_process.call(type, stream)
-      end
-    else
-      Rails.logger.info "Postcode processing ignoring: #{filename}"
-    end
-    block.call(data_summary)
-  end
-
-  def self.process_csv_data(type, csv_stream)
-    return inject_data(csv_stream.read) if type == :dat
-
-    return upsert_csv_data(csv_stream) unless type == :dat
   end
 
   # rubocop:disable Metrics/AbcSize
@@ -510,7 +247,7 @@ CREATE UNIQUE INDEX IF NOT EXISTS os_address_admin_uploads_filename_idx ON os_ad
         p "Processing    #{filename}"
         file_time = Time.current
         read_file("#{directory}/#{filename}", method(:process_csv_data)) do |summation|
-          p "Duration for #{filename}, of #{ActiveSupport::NumberHelper.number_to_human_size summation[:length]} is #{Time.current - file_time}"
+          p "Duration for #{filename}, of #{number_to_human_size summation[:length]} is #{Time.current - file_time}"
           log_postcode_file_loaded(File.basename(filename, File.extname(filename)), summation[:length],
                                    summation[:md5],
                                    summation[:updated_time],
@@ -529,94 +266,7 @@ CREATE UNIQUE INDEX IF NOT EXISTS os_address_admin_uploads_filename_idx ON os_ad
     end
   end
 
-  def self.upsert_csv_data(csv_stream)
-    fully_processed = true
-    csv_headers = os_address_headers.downcase.split(',')
-    SmarterCSV.process(csv_stream, user_provided_headers: csv_headers, chunk_size: 1000, remove_blank_values: false) do |chunk|
-      ActiveRecord::Base.connection_pool.with_connection do |db|
-        db.begin_db_transaction
-        chunk.each do |row|
-          upsert_row(db, row)
-        end
-        db.commit_db_transaction
-      rescue StandardError => e
-        db.rollback_db_transaction
-        p "\tError with upsert: #{e.message}"
-        Rails.logger.error((["POSTCODE: #{e.message}"] + e.backtrace).join($INPUT_RECORD_SEPARATOR))
-        fully_processed = false
-        raise e
-      end
-    end
-    fully_processed
-  end
-  # rubocop:enable Metrics/AbcSize
-
-  INSERT_COLUMNS  = %i[uprn udprn class parent_uprn last_update_date rm_organisation_name sub_building_name building_name building_number
-                      sao_start_number sao_start_suffix sao_end_number sao_end_suffix sao_text
-                      pao_start_number pao_start_suffix pao_end_number pao_end_suffix pao_text street_description dependent_thoroughfare
-                      thoroughfare dependent_locality locality town_name administrative_area post_town postcode postcode_locator po_box_number ward_code].freeze
-  INTEGER_COLUMNS = %i[uprn udprn parent_uprn sao_start_number sao_end_number pao_start_number pao_end_number].freeze
-
-  # rubocop:disable Metrics/CyclomaticComplexity
-  def self.upsert_row(db, row)
-    new_date = DateTime.parse(row[:last_update_date])
-    result   = db.exec_query(os_address_select(row))
-    db_date  = result.empty? ? new_date : DateTime.parse(result[0]['last_update_date'])
-    db.execute(os_address_delete(row)) if db_date < new_date || result.length > 1
-    db.execute(os_address_insert(row)) if result.empty? || new_date > db_date || result.length > 1
-  end
-
-  # rubocop:enable Metrics/CyclomaticComplexity
-
-  def self.os_address_select(row)
-    "select last_update_date from os_address where postcode_locator = '#{row[:postcode_locator]}' and uprn = #{row[:uprn]} order by last_update_date desc"
-  end
-
-  def self.os_address_delete(row)
-    "delete from os_address where postcode_locator = '#{row[:postcode_locator]}' and uprn = #{row[:uprn]}"
-  end
-
-  def self.os_address_insert(row)
-    "insert into os_address (#{INSERT_COLUMNS.map(&:to_s).join(',')}) values (#{INSERT_COLUMNS.map { |c| db_value(row, c) }.join(',')})"
-  end
-
-  def self.db_value(row, col)
-    return 'null' if row[col].nil?
-
-    return "'#{row[col].to_s.gsub("'", "''")}'" unless INTEGER_COLUMNS.include?(col)
-
-    row[col].to_s
-  rescue StandardError => e
-    p (["dbValue Error processing column [:#{col}] in #{row}: #{e.message}"] + e.backtrace).join($INPUT_RECORD_SEPARATOR)
-  end
-
-  def self.inject_data(lines)
-    unless lines.nil?
-      ActiveRecord::Base.connection_pool.with_connection do |conn|
-        rc = conn.raw_connection
-        rc.exec('COPY os_address FROM STDIN WITH CSV')
-        rc.put_copy_data lines
-        rc.put_copy_end
-      end
-    end
-  rescue StandardError => e
-    Rails.logger.error e.message
-  end
-
-  def self.truncate_os_addresses
-    ActiveRecord::Base.connection_pool.with_connection do |db|
-      db.execute('truncate table os_address;')
-    end
-    ActiveRecord::Base.connection_pool.with_connection do |db|
-      db.execute('vacuum os_address;')
-    end
-  rescue StandardError => e
-    p "\tError with truncate: #{e.message}"
-    Rails.logger.error((["POSTCODE truncate: #{e.message}"] + e.backtrace).join($INPUT_RECORD_SEPARATOR))
-    raise e
-  end
-
-  # rubocop:disable CyclomaticComplexity, PerceivedComplexity, Metrics/AbcSize
+  # rubocop:disable CyclomaticComplexity, PerceivedComplexity
   def self.import_postcodes(folder_root, access_key, secret_key, bucket, region)
     OpenURI::Buffer.send :remove_const, 'StringMax' if OpenURI::Buffer.const_defined?('StringMax')
     OpenURI::Buffer.const_set 'StringMax', 0
@@ -641,8 +291,8 @@ CREATE UNIQUE INDEX IF NOT EXISTS os_address_admin_uploads_filename_idx ON os_ad
                                  summation[:updated_time],
                                  DateTime.now.utc)
       end
-    rescue Aws::S3::Errors::AccessDenied => _e
-      Rails.logger.warn "Access denied on #{obj.key}"
+    rescue Aws::S3::Errors::AccessDenied => e
+      Rails.logger.warn "Access denied on #{obj.key}. #{e.message}}"
       puts "*** Access denied on #{obj.key}"
     rescue StandardError => e
       p "\tError: #{e.message}"
